Actualizado Vistas basadas en funciones

+++ profesores/views.py
@@ -21,16 +21,19 @@ def profesores(request):
     else:
         return Response(status=status.HTTP_400_BAD_REQUEST)
 
-@api_view(['DELETE'])
-def profesorId(request):
-    profesor = Profesor.objects.get(id=id)
+@api_view(['GET', 'DELETE'])
+def profesor(request, profesor_id):
+    try:
+        profesor = Profesor.objects.get(id=profesor_id)
+    except Profesor.DoesNotExist:
+        return Response(status=status.HTTP_404_NOT_FOUND)
+
+    if request.method == 'GET':
+        serialized = ProfesorSerializer(profesor)
+        return Response(status=status.HTTP_200_OK, data=serialized.data)
+
     if request.method == 'DELETE':
         profesor.delete()
         return Response(status=status.HTTP_200_OK)
     else:
         return Response({'No se ha podido eliminar el profesor '}, + profesor, + ' ', status=status.HTTP_400_BAD_REQUEST)
-
-
-
-
-
